TheHeader, App basic styles, template + script add
