+++ src/main.js
@@ -2,7 +2,7 @@ import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router.js";
 
-import "./assets/main.css";
+// import "./assets/main.css";
 
 const app = createApp(App);
