+++ src/pages/feedbacks/FeedbackList.vue
@@ -1,0 +1,2 @@
+<template></template>
+<script></script>
