--- src/router/router.js
@@ -1,39 +1,0 @@
-import { createRouter, createWebHistory } from "vue-router";
-
-import FeedbackAdd from './pages/feedbacks/FeedbackAdd.vue';
-import FeedbackDetail from './pages/feedbacks/FeedbackDetail';
-import FeedbackEdit from './pages/feedbacks/FeedbackEdit';
-import FeedbackList from './pages/feedbacks/FeedbackList';
-import StatusFeedbackList from './pages/roadmap/StatusFeedbackList.vue';
-import NotFOund from './pages/NotFound.vue'
-
-const router = createRouter({
-  history: createWebHistory(),
-  routes: [
-    {
-      path: "/", redirect: '/feedbacks'
-    },
-    {
-      path: "/feedbacks",
-      component: FeedbackList,
-    },
-    {
-      path: "/feedbacks/:id",
-      component: FeedbackDetail,
-      children: [
-        {path: 'add', component: FeedbackAdd},
-        {path:'edit', component: FeedbackEdit} //editing feedback
-      ]
-    },
-    {
-      path: "/roadmap",
-      component: StatusFeedbackList,
-    },
-    {
-      path: "/:notFound(.*)",
-      component: NotFOund,
-    }
-  ],
-});
-
-export default router;
